Bounded unsolicited message tx
The unsolicited message tx is a futures::mpsc::UnboudedSender, which means that
if we send unsolicited messages faster than the reader consumes them, the
mpsc will grow until either the reader closes the Connection, or we OOM.
This is often what happens when a server under heavy load is subscribing to
audit messages.

Make the Connection generic over all futures::Sink, keep the UnboundedSender
as the implicit default generic type to avoid breaking compatibility with other
netlink-rs crates, and expose a new public constructor to obtain a Connection
backed by a bounded futures::mpsc::channel.

In the Connection future, if decoded packets are still waiting to be sent to
the Sink, don't read from the socket, we will try again once all packets have
been sent out. Otherwise the Protocol's VecDeque will become the new point of
unconstrained growth.

Note that the requests queue coming from the handle, and the response queue
for each handle request are still unbounded.

The Future implementation should probably be split into two seperate concurrent
aysnc tasks, one for reading from the socket, demuxing and sending to the Sink
or request channels, and one for reading requests from the handle and writing
them to the socket. But unfortunately I don't think this can be done because
unlike tokio's AsyncRead and AsyncWrite, netlink-sys::AsyncSocket can't be
split in two parts, and since it needs to be shared between the two futures,
they wouldn't be able to concurrently sleep on it. So we're stuck with a single
huge future that tries to do everything at the same time.